Various fixes (#559)

+++ src/MetroidPrime/CScriptMailbox.cpp
@@ -22,10 +22,12 @@ CScriptMailbox::CScriptMailbox(CInputStream& in, const CWorldSaveGameInfo& world
     }
   }
 
-  CMemory::OffsetFakeStatics(sizeof(*this));
+  CMemory::OffsetFakeStatics(static_cast< int >(sizeof(*this)));
 }
 
-CScriptMailbox::~CScriptMailbox() { CMemory::OffsetFakeStatics(-sizeof(*this)); }
+CScriptMailbox::~CScriptMailbox() {
+  CMemory::OffsetFakeStatics(-static_cast< int >(sizeof(*this)));
+}
 
 void CScriptMailbox::PutTo(COutputStream& out, CWorldSaveGameInfo& world) const {
   rstl::vector< bool > relayStates(world.GetRelays().size(), false);
@@ -58,7 +60,8 @@ void CScriptMailbox::SendMsgs(const TAreaId& areaId, CStateManager& mgr) {
 
     TEditorId relayId = it->GetRelayId();
     if (HasMsg(relayId)) {
-      mgr.SendScriptMsg(kInvalidUniqueId, targetId, EScriptObjectMessage(it->GetMessage()), kSS_Any);
+      mgr.SendScriptMsg(kInvalidUniqueId, targetId, EScriptObjectMessage(it->GetMessage()),
+                        kSS_Any);
       if (it->GetActive()) {
         hasActiveRelays = true;
       }
